refactor(v3): relocate the PlaybackTrace DTO mappers out of handlers_sessions.go
handlers_sessions.go (1394 LOC) was dominated by DTO mapping that has nothing to
do with its two HTTP handlers. Move the ~590-line PlaybackTrace mapper cluster
(mapSessionPlaybackTrace + mapPlaybackTraceHLSDebug + the mapRuntimePolicy* /
replay sub-mappers + the trace helpers) into a sibling session_playback_trace_mapping.go.

Pure same-package move (no call-site changes, no behavior change):
- handlers_sessions.go 1394 -> 790 (handlers + feedback logic)
- session_playback_trace_mapping.go: the model -> PlaybackTrace DTO mappers

Build + go vet clean; v3 contract/golden tests green; golangci-lint --new-from-rev
= 0; verify-config green.
